// types/movie.ts
// types/movie.ts
export interface Movie {
  Title: string;
  Year?: string;
  imdbID: string;
  Type?: string;
  Poster?: string;
}

export interface MovieDetail {
  Title: string;
  Year: string;
  Rated?: string;
  Released?: string;
  Runtime?: string;
  Genre?: string;
  Director?: string;
  Writer?: string;
  Actors?: string;
  Plot?: string;
  Language?: string;
  Country?: string;
  Awards?: string;
  Poster?: string;
  imdbRating?: string;
  imdbID: string;

  // OMDb response metadata
  Response?: "True" | "False";
  Error?: string;
}